issue #1712: DPI return integer on android.

+++ cocos2dx/platform/android/jni/DPIJni.h
@@ -3,7 +3,7 @@
 
 extern "C" {
 
-float getDPIJNI();
+int getDPIJNI();
 
 } // extern "C"
